remove circular dep
